system: Go back to modified strict state sync method
The original method pre-fork still exhibited some problems, so
the one done earlier in bsnes-jg's life has been reapplied. It is
a futile endeavour to try to get deterministic states when syncing
using libco. Internal runahead does not sync at all for this reason,
but this limits those states to being useful only to the same process
where they were created. To make a long story short, deterministic
states that will work between instances for netplay or runahead
are not happening unless libco is removed.