system: Fix state saving for coprocessor games
When using the coprocessor delayed sync option, state operations
often fail due to the inability for the libco coroutines to sync
correctly to the point where a stable state may be saved. This will
result in becoming stuck in an infinite loop, often with some chips
being emulated while others are stuck, requiring the user to SIGKILL
the process. In order to alleviate this, the coprocessor delayed sync
option is toggled on for the purpose of syncing for states, then
turned back off again if it was off previously. This will not affect
accuracy during play, as it only affects synchronization being done
for the purpose of saving a non-volatile state to disk.